pyproject.toml
[project]
name = "dhen-chal"
version = "0.1.0"
description = "Project is created for Dhen Challenge 2 Hackathon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
]
